Create dual processor
* Remove data and value processors
